upstream: separate binance module
